team11/tripPlanService/requirements.txt
django~=5.2.11
djangorestframework~=3.15.0

psycopg2-binary~=2.9.9
pymongo~=4.8.0

protobuf
grpcio~=1.65.0
grpcio-tools~=1.65.0

celery~=5.4.0
redis~=5.0.0

requests~=2.32.0
python-decouple~=3.8

numpy~=1.26.0
pandas~=2.2.0
scikit-learn~=1.5.0

reportlab~=4.0.0